shop/logistics_views.py

from rest_framework import generics, permissions, status
from rest_framework.response import Response
from rest_framework.views import APIView
from shop.models import Order
from shop.serializers import OrderSerializer

class LogisticsRegisterView(APIView):
    permission_classes = [permissions.IsAuthenticated]

    def post(self, request, order_id):
        try:
            order = Order.objects.get(id=order_id, user=request.user)
        except Order.DoesNotExist:
            return Response({"error": "Order not found."}, status=status.HTTP_404_NOT_FOUND)

        # Simulate logistics registration (e.g., send request to external API)
        logistics_info = {
            "tracking_code": f"TRK-{order.id:06}",
            "carrier": "SampleLogistics",
            "status": "Registered"
        }

        # In a real case, store logistics info in DB or update order model
        return Response({"message": "Logistics registered.", "logistics": logistics_info})
